Add files via upload
As the files are big, the input data can be directly obtained from:
https://www.kaggle.com/donorschoose/io

4 files will have to be downloaded in order for both the codes to work - Donations.csv, Donors.csv, Projects.csv, & Schools.csv

donorschoose-sl.py is the implementation of purely supervised learning using LR, LDA, KNN, & CART algorithms.
donorschoose-usl-sl.py is the implementation of K-Prototypes clustering method in combination with LR, LDA, KNN, & CART algorithms.

The output of the purely supervised approach is attached as donor-predictions-cart.csv. The output of the combination approach is attached as proj-kprototypes-200k-K15.csv showing the result of K-Prototypes method and donor-predictions-usl-sl-cart3.csv showing the final result of the combination technique.